code/requirements.txt
Flask
Jinja2==2.11.3
MarkupSafe==1.1.1
opencv-python
pandas
Pillow
python-dateutil==2.8.1
pytz==2021.1
requests
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
pypylon
prometheus_client
psutil
pydantic
cacheout
python-snap7
pyyaml
pyinstaller
h5py==2.10
pymongo